@@ -8,24 +8,24 @@ repository: https://github.com/LeGoffMael/video_editor
 
 environment:
   flutter: ">=3.0.0"
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  ffmpeg_kit_flutter_full: ^5.1.0
-  path: ^1.8.0 # update to `1.8.1` causes #79
-  path_provider: ^2.0.12
+  ffmpeg_kit_flutter_full: ^6.0.3
+  path: ^1.8.3
+  path_provider: ^2.1.2
   transparent_image: ^2.0.1 # show fade-in placeholder in thumbnails generation
-  video_player: ^2.5.2
+  video_player: ^2.8.3
   video_thumbnail: ^0.5.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.2
 
 screenshots:
   - description: 'Video editor demo'
